fix zmq pipeline ci failures
